Without specifying a data type 'int64' can be used (and is with a newer numpy on a 64-bit platform). Forcing the data type to 'uint8'. (See #2067)
git-svn-id: file:///home/svn/omero/trunk@7784 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/tools/OmeroPy/test/integration/helpers.py
@@ -1,20 +1,24 @@
 import omero.util.script_utils as scriptUtil
-from numpy import arange
+from numpy import arange, uint8
 
-def createTestImage(session):    
+def createTestImage(session):
     gateway = session.createGateway()
     renderingEngine = session.createRenderingEngine()
     queryService = session.getQueryService()
     pixelsService = session.getPixelsService()
     rawPixelStore = session.createRawPixelsStore()
-    
-    plane2D = arange(256).reshape(16,16)
+
+    plane2D = arange(256, dtype=uint8).reshape(16,16)
     pType = plane2D.dtype.name
-    pixelsType = queryService.findByQuery("from PixelsType as p where p.value='%s'" % pType, None) # omero::model::PixelsType
-    image = scriptUtil.createNewImage(pixelsService, rawPixelStore, renderingEngine, pixelsType, gateway, [plane2D], "imageName", "description", dataset=None)
-    
+    # omero::model::PixelsType
+    pixelsType = queryService.findByQuery(
+        "from PixelsType as p where p.value='%s'" % pType, None)
+    image = scriptUtil.createNewImage(pixelsService, rawPixelStore,
+        renderingEngine, pixelsType, gateway, [plane2D],
+        "imageName", "description", dataset=None)
+
     gateway.close()
     renderingEngine.close()
     rawPixelStore.close()
-    
+
     return image.getId().getValue()
